Improve usage on app transformation
