+++ src/App.js
@@ -8,7 +8,7 @@ import RepositoryLink from "./components/RepositoryLink";
 import ThemeCSSVariables from "./components/ThemeCSSVariables";
 import ThemeSwitcher from "./components/ThemeSwitcher";
 import * as theme from "./constants/theme";
-import { logo, repository } from "./helpers/data";
+import { logo, repository } from "./constants/data";
 
 class App extends React.Component {
   static THEME_SESSION_STORAGE_KEY = "APP_THEME";
